+++ src/pybear_dask/model_selection/autogridsearch/_autogridsearch_wrapper/_refit_can_be_skipped.py
@@ -10,7 +10,6 @@ from typing import (
     Callable,
     Literal
 )
-from typing_extensions import Union
 
 from ...GSTCV._GSTCVDask import GSTCVDask
 
@@ -18,7 +17,7 @@ from ...GSTCV._GSTCVDask import GSTCVDask
 
 def _refit_can_be_skipped(
     _GridSearchParent,
-    _scoring: Union[None, str, list, Callable, dict, Literal[False]],
+    _scoring: None | str | list | Callable | dict | Literal[False],
     _total_passes: int
 ) -> bool:
     """Determine if the parent GridSearch, the scoring strategy, and the
@@ -39,7 +38,7 @@ def _refit_can_be_skipped(
     ----------
     _GridSearchParent : GridSearchCV instance
         The parent GridSearchCV class passed to the agscv wrapper.
-    _scoring : Union[None, str, list, Callable, dict, Literal[False]]
+    _scoring : None | str | list | Callable | dict | Literal[False]]
         The value passed to the `scoring` parameter of the parent
         GridSearchCV. On the off chance that the parent GridSearch does
         not have a `scoring` parameter, Literal[False] is passed to here.