+++ tests/model_selection/autogridsearch/verification_AutoGSTCVDask_1min39sec_test.py
@@ -32,7 +32,7 @@ class TestAutoGSTCVDask:
     #         params: ParamsType,
     #         total_passes: int = 5,
     #         total_passes_is_hard: bool = False,
-    #         max_shifts: Union[None, int] = None,
+    #         max_shifts: None | int = None,
     #         agscv_verbose: bool = False,
     #         **parent_gscv_kwargs