+++ tests/model_selection/autogridsearch/verification_AutoGridSearchCVDask_Xsec_test.py
@@ -33,7 +33,7 @@ class TestAutoGridSearchDask:
     # params: ParamsType,
     # total_passes: int = 5,
     # total_passes_is_hard: bool = False,
-    # max_shifts: Union[None, int] = None,
+    # max_shifts: None | int = None,
     # agscv_verbose: bool = False,
     # **parent_gscv_kwargs